requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
openai>=1.12.0
google-cloud-vision>=3.5.0
google-auth>=2.27.0
gspread>=6.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.26.0
openpyxl>=3.1.0
pdf2image>=1.16.0
Pillow>=10.0.0
watchdog>=3.0.0
customtkinter>=5.2.0
psutil>=5.9.0
requests>=2.31.0
